+++ tests/testthat/test-CB2FindCell.R
@@ -1,31 +1,28 @@
 data(mbrainSub)
-CBOut <- CB2FindCell(mbrainSub, FDR_threshold = 0.01,
-                    lower = 100, Ncores = 2,PrintProg = FALSE)
 
-
-test_that("Output CBout format", {
-    expect_true(all(c("cluster_matrix","cell_matrix")%in%names(CBOut)))
-})
-
-test_that("Background threshold", {
-    for(pooling in c(50, 100)){
+for(lower in c(50, 100)){
+    for(upper in c(500, 1000)){
         CBOut_temp <- CB2FindCell(mbrainSub,
-                            lower = pooling, 
-                            Ncores = 2, PrintProg = FALSE)
-        expect_true(all(Matrix::colSums(CBOut_temp$cluster_matrix)>pooling))
-        expect_true(all(Matrix::colSums(CBOut_temp$cell_matrix)>pooling))
+                                lower = lower, 
+                                upper = upper,
+                                Ncores = 2, verbose = FALSE)
+        
+        test_that("Output format", {
+            expect_true(all(c("cluster_matrix","cell_matrix")%in%names(CBOut_temp)))
+        })
+        
+        cell_b <- colnames(cbind(CBOut_temp$cluster_matrix,
+                                 CBOut_temp$cell_matrix))
+    
+        test_that("Lower threshold", {
+            expect_true(all(Matrix::colSums(mbrainSub[,cell_b])>lower))
+        })
+    
+        upper_b <- colnames(mbrainSub)[Matrix::colSums(mbrainSub)>=upper]
+        test_that("upper threshold", {
+            expect_true(all(upper_b%in%cell_b))
+        })
+        
     }
-})
-                                                                                
+}
 
-test_that("Upper threshold", {
-    for(Upper in c(500, 1000)){
-        CBOut_temp <- CB2FindCell(mbrainSub,
-                            upper = Upper,
-                            Ncores = 2, PrintProg = FALSE)
-        Upper_b <- colnames(mbrainSub)[Matrix::colSums(mbrainSub)>=Upper]
-        Cell_b <- colnames(cbind(CBOut_temp$cluster_matrix,
-                            CBOut_temp$cell_matrix))
-        expect_true(all(Upper_b%in%Cell_b))
-    }
-})